Add: loadImage method and Utils class.
Utils class will hold all kind of different utility methods for running
the game engine. loadImage tries to load the image with the given file
name and if it fails, it will print out an error message to console. To
keep the game from crashing it will then return an error image which is
very obvious visually.