Add and implement Jackson for parsing json.
Add jackson for parsing json. It reads the json file and turns it to an
object. For now it is used to create a JacksonMap which is an object that
has the data of a tile maps json file.

Implement the use of JacksonMap for the game engine. Remove the old way of
generating the maps.

Refactor the structure of the project a bit. Createa an util package and
move the Util class within it. Create a json package within the util
package and move JacksonMap within it. Make MapLayer and TileSet inside
JacksonMap public and therefor move them to their own files within the
newly created json package.